neuro_engine/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
kafka-python==3.0.0
kubernetes==29.0.0
prometheus-client==0.19.0
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
cryptography==41.0.3
certifi==2023.11.17